# openapi.json
{
  "components": {
    "schemas": {
...
      "ConfirmRequest": {
        "additionalProperties": false,
        "properties": {
          "corrections": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "description": "Operator-corrected field values keyed by field name. Omitted fields keep the extracted value. Sensitive numbers are hashed and reduced to last4 before persistence.",
            "maxProperties": 64,
            "title": "Corrections",
            "type": "object"
          }
        },
        "title": "ConfirmRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScanListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScanSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "ScanListResponse",
        "type": "object"
      },
      "ScanRequest": {
        "additionalProperties": false,
        "properties": {
          "back_image_base64": {
            "anyOf": [
              {
                "maxLength": 12582912,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "JPEG or PNG",
            "title": "Back Image Base64"
          },
          "country": {
            "description": "ISO 3166-1 alpha-2 or alpha-3 code",
            "pattern": "^[A-Z]{2,3}$",
            "title": "Country",
            "type": "string"
          },
          "document_type": {
            "enum": [
              "national_id",
              "passport",
              "driver_license",
              "residence_permit"
            ],
            "title": "Document Type",
            "type": "string"
          },
          "front_image_base64": {
            "description": "JPEG or PNG",
            "maxLength": 12582912,
            "minLength": 1,
            "title": "Front Image Base64",
            "type": "string"
          }
        },
        "required": [
          "document_type",
          "country",
          "front_image_base64"
        ],
        "title": "ScanRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "WorkloadToken": {
        "description": "Middleware V3 workload token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Standalone multi-tenant document intelligence API. Reachable only via Caddy -> Kong -> Middleware V3 -> Document Intelligence. OCR runs in Codestra-OCR-Workers.\n\n**OCR/QR output is an intake aid, not government authenticity verification.** Raw document images are never persisted; sensitive document numbers are stored only as a tenant-scoped keyed hash plus last four characters.",
    "title": "Codestra Document Intelligence API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/documents": {
      "get": {
        "operationId": "list_documents_v1_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^dscan_[A-Za-z0-9_-]{8,128}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "description": "Required. Calling workload name (allowlisted, e.g. middleware-v3).",
            "in": "header",
            "name": "x-codestra-workload",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required. Calling workload name (allowlisted, e.g. middleware-v3).",
              "title": "X-Codestra-Workload"
            }
          },
          {
            "description": "Required. Tenant resolved by Middleware V3; the only tenant source.",
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required. Tenant resolved by Middleware V3; the only tenant source.",
              "title": "X-Tenant-Id"
            }
          },
          {
            "description": "Optional opaque operator reference.",
            "in": "header",
            "name": "x-codestra-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional opaque operator reference.",
              "title": "X-Codestra-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or malformed tenant identity"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid workload token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Calling workload not allowlisted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WorkloadToken": []
          }
        ],
        "summary": "List recent scan sessions for the authenticated tenant",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/scan": {
      "post": {
        "operationId": "scan_document_v1_documents_scan_post",
        "parameters": [
          {
            "description": "Optional. Replaying the same key with the same images returns the original scan (200).",
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[A-Za-z0-9._:-]{8,128}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional. Replaying the same key with the same images returns the original scan (200).",
              "title": "Idempotency-Key"
            }
          },
          {
            "description": "Required. Calling workload name (allowlisted, e.g. middleware-v3).",
            "in": "header",
            "name": "x-codestra-workload",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required. Calling workload name (allowlisted, e.g. middleware-v3).",
              "title": "X-Codestra-Workload"
            }
          },
          {
            "description": "Required. Tenant resolved by Middleware V3; the only tenant source.",
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required. Tenant resolved by Middleware V3; the only tenant source.",
              "title": "X-Tenant-Id"
            }
          },
          {
            "description": "Optional opaque operator reference.",
            "in": "header",
            "name": "x-codestra-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional opaque operator reference.",
              "title": "X-Codestra-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            },
            "description": "Idempotent replay of an earlier scan"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or malformed tenant identity"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid workload token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Calling workload not allowlisted"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Idempotency-Key reused for a different request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or image"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "OCR worker schema mismatch or error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "OCR worker unavailable"
          }
        },
        "security": [
          {
            "WorkloadToken": []
          }
        ],
        "summary": "Scan a document (front/back images) via the OCR worker",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{scan_id}": {
      "get": {
        "operationId": "get_document_v1_documents__scan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          },
          {
            "description": "Required. Calling workload name (allowlisted, e.g. middleware-v3).",
            "in": "header",
            "name": "x-codestra-workload",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required. Calling workload name (allowlisted, e.g. middleware-v3).",
              "title": "X-Codestra-Workload"
            }
          },
          {
            "description": "Required. Tenant resolved by Middleware V3; the only tenant source.",
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required. Tenant resolved by Middleware V3; the only tenant source.",
              "title": "X-Tenant-Id"
            }
          },
          {
            "description": "Optional opaque operator reference.",
            "in": "header",
            "name": "x-codestra-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional opaque operator reference.",
              "title": "X-Codestra-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or malformed tenant identity"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid workload token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Calling workload not allowlisted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Scan not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "WorkloadToken": []
          }
        ],
        "summary": "Get a scan session (tenant-scoped)",
        "tags": [
          "documents"
        ]
      }
    },
    "/v1/documents/{scan_id}/confirm": {
      "post": {
        "operationId": "confirm_document_v1_documents__scan_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          },
          {
            "description": "Required. Calling workload name (allowlisted, e.g. middleware-v3).",
            "in": "header",
            "name": "x-codestra-workload",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required. Calling workload name (allowlisted, e.g. middleware-v3).",
              "title": "X-Codestra-Workload"
            }
          },
          {
            "description": "Required. Tenant resolved by Middleware V3; the only tenant source.",
            "in": "header",
            "name": "x-tenant-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Required. Tenant resolved by Middleware V3; the only tenant source.",
              "title": "X-Tenant-Id"
            }
          },
          {
            "description": "Optional opaque operator reference.",
            "in": "header",
            "name": "x-codestra-actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional opaque operator reference.",
              "title": "X-Codestra-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or malformed tenant identity"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid workload token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Calling workload not allowlisted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Scan not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Scan not in pending_review"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid corrections"
          }
        },
        "security": [
          {
            "WorkloadToken": []
          }
        ],
        "summary": "Confirm a scan with operator corrections",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}